Add files via upload
Working on custom video player in reactjs. The react player does work although having trouble with the seekbar, so currently building out an audio player to see if I can get the seekbar to work, while the video plays. Was able to get the scroll to work in a different version, but the seek bar and time didn't update. Going to upload audio version and see if I can mix and match these projects to get a fully functional customized react player. [Wanted to avoid generic upload of YT,Vimeo,Twitch etc]